Simplify collected comments and relocate shared helpers.
Show comment counts as display-only, drop unused enter animations, and move collected helpers into lib/types.
