+++ lib/stats/formatCollectedStatValue.ts
